+++ doc.go
@@ -1,0 +1,6 @@
+// slydepay_lib project doc.go
+
+/*
+slydepay_lib document
+*/
+package slydepay_lib
