+++ orderitem.go
@@ -1,0 +1,10 @@
+// slydepay_lib project orderitem
+package slydepay_lib
+
+type OrderItem struct {
+	ItemCode  string
+	ItemName  string
+	Quantity  int16
+	UnitPrice float64
+	SubTotal  float64
+}
