--- slydepay_lib/doc.go
@@ -1,6 +1,0 @@
-// slydepay_lib project doc.go
-
-/*
-slydepay_lib document
-*/
-package slydepay_lib
